% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{adv.tzh-package}
\alias{adv.tzh}
\alias{adv.tzh-package}
\title{adv.tzh: Advanced Manipulation of R Packages}
\description{
To perform advanced manipulation of R packages.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
