% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRelease.R
\name{checkRelease}
\alias{checkRelease}
\title{Additional Checks Before \link[devtools]{release}}
\usage{
checkRelease(pkg, to_release = TRUE)
}
\arguments{
\item{pkg}{\link[base]{character} scalar, directory of the package}

\item{to_release}{\link[base]{logical} scalar}
}
\description{
additional checks before \link[devtools]{release}
}
