% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_parent_dir.R
\name{clean_parent_dir}
\alias{clean_parent_dir}
\title{Clean Augment Files in Parent Directory}
\usage{
clean_parent_dir(pkg = ".")
}
\arguments{
\item{pkg}{..}
}
\description{
Clean Augment Files in Parent Directory
}
\examples{
# clean_parent_dir(pkg = '../groupedHyperframe')
}
