% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackage.R
\name{createPackage}
\alias{createPackage}
\title{Create Sub-Package}
\usage{
createPackage(name, path = tempdir(), ...)
}
\arguments{
\item{name}{\link[base]{character} scalar, name of the package}

\item{path}{\link[base]{character} scalar, parent directory of the package,
default is the return of \link[base]{tempdir}}

\item{...}{additional parameters of \link{package_skeleton_ext}}
}
\description{
Create sub-packages from my master package \strong{tzh}
}
\details{
This function can only be run by developer (i.e., me) from my master package \strong{tzh}.

\link[Rcpp]{Rcpp.package.skeleton} is next step.
}
\references{
\url{https://cran.r-project.org/web/packages/policies.html}

My packages
\url{https://cran.r-project.org/package=QuantileGH}

\url{https://cran.r-project.org/package=DanielBiostatistics10th}

\url{https://cran.r-project.org/package=DemographicTable}

\url{https://cran.r-project.org/package=ThomasJeffersonUniv}
}
