% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFromPackage.R
\name{dataFromPackage}
\alias{dataFromPackage}
\title{Data From Package}
\usage{
dataFromPackage(pkg = "datasets", what)
}
\arguments{
\item{pkg}{\link[base]{character} scalar or \link[base]{vector}, default \code{'datasets'}}

\item{what}{\link[base]{character} scalar or \link[base]{vector}, class(es), see \link[base]{inherits}}
}
\value{
Function \link{dataFromPackage} returns a \link[base]{list} of R objects.
}
\description{
Obtain all objects available via \link[utils]{data} from a package
}
\examples{
# dataFromPackage(c('datasets', 'MASS'), what = 'data.frame') # next step

lapply(dataFromPackage('datasets', what = 'data.frame'), head)
lapply(dataFromPackage('datasets', what = 'numeric'), head)

}
\references{
\url{https://stackoverflow.com/questions/27709936/get-a-list-of-the-data-sets-in-a-particular-package}
}
