% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definedLocal.R
\name{definedLocal}
\alias{definedLocal}
\title{Functions in a Package, defined in a \link[base]{local}}
\usage{
definedLocal(pkg, ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{definedLocal} returns an \link[base]{invisible} \link[base]{numeric} \link[base]{vector}.
}
\description{
Functions in a Package, defined in a \link[base]{local}
}
\examples{
# ?devtools::check error, but works for end-user
\dontrun{definedLocal('spatstat.explore')}
\dontrun{(definedLocal('spatstat.explore'))}
}
