% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_function.R
\name{.local_obj}
\alias{.local_obj}
\title{Objects in a Local Environment}
\usage{
.local_obj(fun)
}
\arguments{
\item{fun}{\link[base]{function}, must be given in the format of \verb{pkg::function}}
}
\value{
Function \code{\link[=.local_obj]{.local_obj()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Objects in a Local Environment
}
\details{
Function \code{\link[=.local_obj]{.local_obj()}} provides the names of other objects
in the \link[base]{local} environment of \code{fun}ction definition.
}
\examples{
.local_obj(fun = base::sum)
.local_obj(fun = base::sub)
.local_obj(fun = stats:::update.packageStatus) # actually ?utils:::update.packageStatus
.local_obj(fun = base::.doSortWrap)
}
\keyword{internal}
