% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_book.R
\name{fancymethods}
\alias{fancymethods}
\alias{methods2kable}
\alias{methods2reactable}
\title{\code{S3} methods table in vignette}
\usage{
fancymethods(
  generic.function,
  class,
  package,
  package_pattern,
  backtick = TRUE,
  include_non_exported = TRUE,
  ...
)

methods2kable(...)

methods2reactable(..., subset, select, backtick = FALSE)
}
\arguments{
\item{generic.function, class}{see function \link[utils]{methods}}

\item{package}{\link[base]{character} scalar}

\item{package_pattern}{\link[base]{character} scalar of \link[base]{regex}}

\item{backtick}{\link[base]{logical} scalar, whether to put backticks around function names.
Default \code{TRUE} for Markdown/Quarto rendering.}

\item{...}{additional parameters of the function \link[utils]{methods}}

\item{subset, select}{see function \link[base]{subset.data.frame}}
}
\value{
The function \code{\link[=methods2kable]{methods2kable()}} returns a \link[base]{data.frame}.
}
\description{
\code{S3} methods table in vignette
}
\keyword{internal}
