% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mtime.R
\name{file_mtime}
\alias{file_mtime}
\title{File Modification Time, Reverse Chronical}
\usage{
file_mtime(
  path,
  pattern = "\\\\.xlsx$|\\\\.xls$|\\\\.csv$",
  file = list.files(path = path, pattern = pattern, ..., full.names = TRUE),
  ...
)
}
\arguments{
\item{path}{\link[base]{character} scalar, directory on hard drive, see \link[base]{list.files}}

\item{pattern}{\link[base]{character} scalar, regular expression \link[base]{regex}}

\item{file}{(optional) \link[base]{character} scalar or \link[base]{vector}}

\item{...}{..}
}
\description{
..
}
\examples{
\dontrun{ # devtools::check error
file_mtime('./R', pattern = '\\\\.R$')
}

}
\seealso{
\link[base]{file.info} \link[base]{file.mtime}
}
\keyword{internal}
