% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funFromPackage.R
\name{funFromPackage}
\alias{funFromPackage}
\alias{funFromThisPackage}
\title{Functions From a Package}
\usage{
funFromPackage(pkg, all.names = TRUE, ...)

funFromThisPackage(all.names = TRUE, ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar, package name}

\item{all.names}{\link[base]{logical} scalar, see \link[base]{ls}.
Default \code{TRUE} to get functions with name starting with \code{'.'}}

\item{...}{additional parameters of \link[base]{ls}}
}
\value{
Functions \link{funFromPackage} and \link{funFromThisPackage}
both return a \link[base]{character} \link[base]{vector} of function names.
}
\description{
Obtain all \link[base]{function}s defined (\emph{not} imported) in a package
}
\details{
Function \link{funFromPackage} ..

Function \link{funFromThisPackage} ..
}
\examples{
funFromPackage('stats', pattern = '\\\\.test$')

}
\references{
\url{https://stackoverflow.com/questions/8696158/find-all-functions-including-private-in-a-package}
}
