% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.missing.packages.R
\name{install.missing.packages}
\alias{install.missing.packages}
\title{Install Missing Packages}
\usage{
install.missing.packages(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}, package name(s)}
}
\description{
..
}
\keyword{internal}
