% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_function.R
\name{local_function}
\alias{local_function}
\title{Functions in a Package, defined inside \link[base]{local}}
\usage{
local_function(pkg, ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{local_function} returns an \link[base]{invisible} \link[base]{numeric} \link[base]{vector}.
}
\description{
Functions in a Package, defined inside \link[base]{local}
}
\note{
Need to \strong{Suggests} for examples to run.
}
\examples{
local_function('base')
local_function('stats')
local_function('utils')
local_function('tools')
local_function('MASS')
local_function('nlme')
local_function('survival')
}
