% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextCRANversion.R
\name{nextCRANversion}
\alias{nextCRANversion}
\title{nextCRANversion}
\usage{
nextCRANversion(name, Version, repos = "https://cran.rstudio.com/")
}
\arguments{
\item{name}{\link[base]{character} scalar, name of package}

\item{Version}{(optional) \link[base]{character} scalar, package version}

\item{repos}{\link[base]{character} scalar, see \link[utils]{contrib.url}.
Default \code{'https://cran.rstudio.com/'},
which is the return of \code{getOption('repos')} when using RStudio}
}
\value{
Function \link{nextCRANversion} returns a \link[base]{numeric_version} object.
}
\description{
..
}
\details{
If package \pkg{name} is not available on CRAN, i.e., this is the first release,
then \code{'0.1.0'} is returned.

Otherwise, if \code{Version} is missing,
a patch to the current CRAN version is created, via \link{patchVersion}.

Otherwise, if user-provided \code{Version} is older than CRAN version,
then the patched CRAN version, as described previously, is returned.

Otherwise, the user-provided \code{Version} is returned.
}
\examples{
nextCRANversion('MASS')
nextCRANversion('MASS', Version = '3.0')
nextCRANversion('MASS', Version = '8.0')

}
