% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectFrom.R
\name{objectFrom}
\alias{objectFrom}
\alias{objectFrom.environment}
\alias{objectFrom.character}
\title{Functions From a Package}
\usage{
objectFrom(x, ...)

\method{objectFrom}{environment}(x, what, all.names = TRUE, sorted = TRUE, ...)

\method{objectFrom}{character}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{what}{parameter of function \link[base]{inherits}}

\item{all.names, sorted, ...}{additional parameters of function \link[base]{ls}}
}
\value{
Functions \code{\link[=objectFrom]{objectFrom()}} returns a \link[base]{character} \link[base]{vector} of function names.
}
\description{
Obtain all \link[base]{function}s defined (\emph{not} imported) in a package
}
\details{
Function \code{\link[=objectFrom]{objectFrom()}} ..
}
\examples{
objectFrom('stats', pattern = '\\\\.test$')
objectFrom('stats', what = 'selfStart')
# objectFrom('.') # for developer only

}
\references{
\url{https://stackoverflow.com/questions/8696158/find-all-functions-including-private-in-a-package}
}
