% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{package_dependencies_date}
\alias{package_dependencies_date}
\alias{working_package_dependencies_date}
\title{Last Update Dates of Packages Dependencies, in Reverse Order}
\usage{
package_dependencies_date(...)

working_package_dependencies_date(...)
}
\arguments{
\item{...}{parameters of \link{package_dependencies_ext} or \link{working_package_dependencies_ext}}
}
\description{
..
}
