% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_skeleton_ext.R
\name{package_skeleton_ext}
\alias{package_skeleton_ext}
\title{Extend \link[utils]{package.skeleton}}
\usage{
package_skeleton_ext(
  name,
  path,
  code_files = stop(""),
  cpp_files = character(),
  rda_files = character(),
  object2rda = character(),
  vignette_files = character(),
  Title = stop("package `Title` missing"),
  Version = "0.1.0",
  Description = stop(""),
  AuthorsR = c(person(given = "Tingting", family = "Zhan", role = c("aut", "cre", "cph"),
    email = c("tingtingzhan@gmail.com", "Tingting.Zhan@jefferson.edu"), comment = c(ORCID
    = "0000-0001-9971-4844"))),
  Depends = character(),
  Imports = character(),
  Suggests = character(),
  ...
)
}
\arguments{
\item{name, path, code_files}{see \link[utils]{package.skeleton}}

\item{cpp_files}{next step, using \code{Rcpp::Rcpp.package.skeleton}}

\item{rda_files}{(optional) \link[base]{character} \link[base]{vector}}

\item{object2rda}{(optional) \link[base]{character} \link[base]{vector}, name(s) of objects to be
saved in an \code{.rda} file.}

\item{vignette_files}{\link[base]{character} \link[base]{vector}}

\item{Title}{\link[base]{character} scalar}

\item{Version}{(optional) \link[base]{character} scalar}

\item{Description}{\link[base]{character} scalar}

\item{AuthorsR}{a \link[utils]{person} \link[base]{call}.
The \code{'Author'} and \code{'Maintainer'} fields in DESCRIPTION file will be generated \strong{by CRAN} from this argument}

\item{Depends}{\link[base]{character} \link[base]{vector},
names of packages}

\item{Imports}{\link[base]{character} \link[base]{vector},
names of packages}

\item{Suggests}{\link[base]{character} \link[base]{vector},
names of packages}

\item{...}{..}
}
\description{
Extend \link[utils]{package.skeleton}
}
\details{
This function can only be run by developer (i.e., me) from my master package \strong{tzh}.
}
\references{
\url{https://cran.r-project.org/web/packages/policies.html}
}
