% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{packageAdvanced-package}
\alias{packageAdvanced}
\alias{packageAdvanced-package}
\title{packageAdvanced}
\description{
packageAdvanced
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} [copyright holder]

}
\keyword{internal}
