% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageIQR_ext.R
\name{packageIQR_S3}
\alias{packageIQR_S3}
\alias{as.environment.packageIQR}
\alias{split.packageIQR}
\title{S3 Method Dispatches for \code{'packageIQR'} Object}
\usage{
\method{as.environment}{packageIQR}(x)

\method{split}{packageIQR}(x, ...)
}
\arguments{
\item{x}{\code{'packageIQR'} object, returned value of function \link[utils]{data},
all available data sets in one or more packages}

\item{...}{addtional parameters, currently not in use}
}
\value{
Function \code{\link[=as.environment.packageIQR]{as.environment.packageIQR()}} returns an \link[base]{environment}.
}
\description{
S3 Method Dispatches for \code{'packageIQR'} Object
}
\details{
Function \code{\link[=as.environment.packageIQR]{as.environment.packageIQR()}} re-runs function \link[utils]{data}
in a \link[base]{new.env}.

Function \code{\link[=split.packageIQR]{split.packageIQR()}} is inspired by function \code{?utils:::print.packageIQR}.
}
\note{
For data name without parentheses, e.g. \code{CO2},
\itemize{
\item {works: \code{data(list = 'CO2', package = 'datasets')}}
\item {works: \code{mget('CO2', envir = as.environment('package:datasets'))}}
}

Data names with parentheses, e.g. \code{beaver1 (beavers)},
are objects saved to \code{beavers.rda} file in the source.
\itemize{
\item {works: \code{data(list = 'beavers', package = 'datasets')}}
\item {no! \code{data(list = 'beaver1', package = 'datasets')}}
\item {no! \code{mget(x = 'beavers', envir = as.environment('package:datasets'))}}
\item {works: \code{mget(x = 'beaver1', envir = as.environment('package:datasets'))}}
}

\link[base]{mget} is ~200 times faster than \link[utils]{data}.

Sometimes function \link[base]{mget} fails and I do not know why.

Function \link[utils]{data} looks for 'data' in \emph{many} places.
}
\examples{
ls1 = data(package = 'datasets') |> as.environment() |> as.list(sort = TRUE)
ls2 = as.environment('package:datasets') |> as.list(sort = TRUE)
stopifnot(identical(ls1, ls2))

if (FALSE) {
ls1 = data(package = 'spatstat.data') |> as.environment() |> as.list(sort = TRUE)
library(spatstat.data); ls2 = as.environment('package:spatstat.data') |> as.list(sort = TRUE)
length(ls1)
length(ls2)
setdiff(names(ls2), names(ls1))
class(ls2$copyExampleFiles) # function
}

\dontrun{
pkg = setdiff(rownames(installed.packages()), c('rjags', 'VennDiagram'))
ev = data(package = pkg) |> as.environment() # not that slow
length(ev)
}
x = data(package = c('datasets', 'survival'))
x1 = data(package = c('datasets'))
x2 = data(package = c('survival'))
y = split(x)
stopifnot(identical(y$datasets, x1), identical(y$survival, x2))
}
