% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_.R
\name{patch_}
\alias{patch_}
\title{Create a Patch}
\usage{
patch_(v)
}
\arguments{
\item{v}{\link[base]{character} scalar, or \link[base]{numeric_version} object,
package version}
}
\value{
Function \code{\link[=patch_]{patch_()}} returns a \link[base]{numeric_version} object.
}
\description{
Create a \emph{patch} to the input version.
}
\details{
A \emph{patch} to the input version is created. For example,
\itemize{
\item \code{'2.3.1'} is returned if the input version is \code{'2.3'} or \code{'2.3.0'}

\item \code{'2.3.2'} is returned if the input version is \code{'2.3.1'}
}
}
\examples{
patch_('7.3.6')
'MASS' |> packageVersion() |> patch_()
'ordinal' |> packageVersion() |> patch_()
}
\references{
\url{https://r-pkgs.org/release.html}
}
