% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackage.R
\name{patchVersion}
\alias{patchVersion}
\title{Create a Patch to Current CRAN Version}
\usage{
patchVersion(v)
}
\arguments{
\item{v}{\link[base]{character} scalar, or \link[base]{numeric_version} object,
package version}
}
\value{
\link{patchVersion} returns a \link[base]{numeric_version} object.
}
\description{
Create a \emph{patch} to the current CRAN Version.
}
\details{
A \emph{patch} to the current CRAN version is created. For example,
\itemize{
\item \code{'2.3.1'} is returned if the current CRAN version is \code{'2.3'} or \code{'2.3.0'}

\item \code{'2.3.2'} is returned if the current CRAN version is \code{'2.3.1'}
}
}
\examples{
patchVersion('7.3.6')
patchVersion(packageVersion('MASS'))

}
\references{
\url{https://r-pkgs.org/release.html}
}
