% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_cleanup.R
\name{pkg_cleanup}
\alias{pkg_cleanup}
\title{Development Package Clean Up}
\usage{
pkg_cleanup(pkg = ".", rm_o = FALSE)
}
\arguments{
\item{pkg}{\link[base]{character} scalar}

\item{rm_o}{\link[base]{logical} scalar, whether to remove the \verb{./src/*.o} files.
Default \code{FALSE}.}
}
\description{
..
}
\details{
Function \link{pkg_cleanup} prepares the working package \code{'.'} for compilation
\itemize{
\item {removes unrecognized files from main directory}
\item {removes size-0 \verb{./R/*.R} and \verb{./src/*.cpp} files}
\item {updates the \emph{Date}, \emph{Version} in DESCRIPTION file.
Always put S4 \link[methods]{setClass} and \link[methods]{setMethod}
in a same file, then we do not need \emph{Collate} in DESCRIPTION file}
}
}
\examples{
if (FALSE) {
  pkg_cleanup('~/Dropbox/Packages/tzh') 
}
# # only in devtools::load_all('.') mode!!

}
