% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_installed.R
\name{pkg_installed}
\alias{pkg_installed}
\title{Check if Packages are Installed}
\usage{
pkg_installed(x)
}
\arguments{
\item{x}{\link[base]{character} scalar or \link[base]{vector}, package name(s)}
}
\description{
..
}
