% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_latest.R
\name{pkg_latest}
\alias{pkg_latest}
\title{Latest Updated Packages}
\usage{
pkg_latest(within_days = 5L)
}
\arguments{
\item{within_days}{\link[base]{integer} scalar,
number of days to look back (default \code{5L})}
}
\description{
..
}
\details{
Function \link{pkg_latest} returns the date of update for the most recently updated packages.
}
\examples{
pkg_latest()

}
