% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_book.R
\name{pkgInfo.}
\alias{pkgInfo.}
\title{Package Info}
\usage{
pkgInfo.(package)
}
\arguments{
\item{package}{\link[base]{character} scalar}
}
\description{
Package Info
}
\note{
Potential name clash \link[sessioninfo]{package_info}.
}
