% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_function.R
\name{print.local_obj}
\alias{print.local_obj}
\title{print.local_obj}
\usage{
\method{print}{local_obj}(x, details = FALSE, ...)
}
\arguments{
\item{x}{returned object from function \code{\link[=.local_obj]{.local_obj()}}}

\item{details}{\link[base]{logical} scalar, whether to print all local objects.
Default \code{FALSE}.}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=print.local_obj]{print.local_obj()}} prints an ANSI string,
see \link[cli]{ansi-styles},
but does not have a returned value.
}
\description{
print.local_obj
}
\examples{
.local_obj(fun = base::.doSortWrap) |> 
  print(details = TRUE)
}
\keyword{internal}
