% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxed_identical.R
\name{relaxed_identical}
\alias{relaxed_identical}
\title{Relaxed Identical}
\usage{
relaxed_identical(x, y)
}
\arguments{
\item{x, y}{any \strong{R} objects}
}
\value{
Function \link{relaxed_identical} returns a \link[base]{logical} scalar.
}
\description{
Slightly relax the \link[base]{identical} criteria,
to identify near-identical objects,
e.g., model estimates with mathematically identical model specifications.
}
\details{
Function \code{\link[=relaxed_identical]{relaxed_identical()}} relaxes function \link[base]{identical} in the following ways.
\subsection{For object with \link[base]{typeof} being \link[base]{double} and/or \link[base]{integer}}{

Test near equality with \link[base]{attributes} ignored,
i.e., function \link[base]{all.equal.numeric} with option \code{check.attributes = FALSE}.
}

\subsection{For \link[stats]{formula}s}{

Set \link[base]{environment} of \code{x} and \code{y} to \code{NULL},
then compare them using function \link[base]{identical}.
Note that
\itemize{
\item{\link[stats]{formula} \link[base]{is.recursive}, thus must be placed before the \link[base]{is.recursive} branching;}
\item{\link[stats]{formula} is \emph{not} \link[base]{closure}.
Therefore, using function \link[base]{identical} with option \code{ignore.environment = TRUE} does \emph{not} work!}
}
}

\subsection{For \link[base]{function}s}{

Ignore \link[base]{environment} of \code{x} and \code{y},
i.e., using function \link[base]{identical} with option \code{ignore.environment = TRUE}.
Note that
\itemize{
\item{\link[base]{function} \link[base]{is.recursive}, whether it is \link[base]{closure} or \link[base]{is.primitive}.
Therefore it must be placed before the \link[base]{is.recursive} branching;}
}
}

\subsection{For all other \link[base]{is.recursive} objects}{

Function \code{\link[=relaxed_identical]{relaxed_identical()}} is called \emph{\strong{recursively}}, for each \link[base]{$} element of \code{x} and \code{y}.
}

\subsection{For \link[base]{S4} objects}{

Function \link{relaxed_identical} is called \emph{\strong{recursively}}, for each \link[base]{@} \link[methods]{slot}
(which is technically the \link[base]{attributes}) of \code{x} and \code{y},
including the \verb{@.Data} slot.
Note that
\itemize{
\item \link[base]{S4} objects are \emph{\strong{not}} \link[base]{is.recursive}.
}
}

\subsection{Otherwise}{

Function \link[base]{identical} is called, as the exception handling.
}
}
\examples{
# mathematically identical model specification
m1 = lm(breaks ~ -1 + wool + wool:tension, data = warpbreaks)
m2 = lm(breaks ~ -1 + tension + tension:wool, data = warpbreaks)
foo = function(m) list(pred = predict(m), resid = residuals(m))
identical(foo(m1), foo(m2)) # FALSE
stopifnot(relaxed_identical(foo(m1), foo(m2)))
}
