% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.envir.R
\name{save.envir}
\alias{save.envir}
\title{Save an \link[base]{environment} to \code{.rda} File}
\usage{
save.envir(envir = parent.frame(), compress = "xz", ...)
}
\arguments{
\item{envir}{an R object convertible to \link[base]{environment}}

\item{compress}{\link[base]{character} scalar, see \link[base]{save}.
Default \code{'xz'}}

\item{...}{additional parameters of \link[base]{save}, other than
\code{list}, \code{envir} and \code{compress}}
}
\description{
..
}
