% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{sort_packageDate_}
\alias{sort_packageDate_}
\title{Last Update Dates of Packages, in Reverse Order}
\usage{
sort_packageDate_(pkg, decreasing = TRUE, ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar or \link[base]{vector},
packages on CRAN}

\item{decreasing}{\link[base]{logical} scalar, default \code{TRUE}.
See \link[base]{order}}

\item{...}{additional parameters of \link[utils]{packageDate}}
}
\value{
Function \code{\link[=sort_packageDate_]{sort_packageDate_()}} returns a \link[base]{Date} scalar or \link[base]{vector}.
}
\description{
..
}
\examples{
# To pass CRAN check
'https://cran.rstudio.com/' |>
 utils::contrib.url() |> 
 utils::available.packages() |> 
 tools::package_dependencies(c('grid', 'pkgload'), db = _, recursive = TRUE) |>
 lapply(FUN = sort_packageDate_)

\dontrun{# In RStudio, simply use
tools::package_dependencies(c('grid', 'pkgload'), recursive = TRUE) |>
 lapply(FUN = sort_packageDate_)
}
}
\keyword{internal}
