% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-.MethodsFunction.R
\name{sub_MethodsFunction}
\alias{sub_MethodsFunction}
\alias{[.MethodsFunction}
\title{Subset of S3/S4 Methods}
\usage{
\method{[}{MethodsFunction}(x, i)
}
\arguments{
\item{x}{a \code{'MethodsFunction'} object, returned value from functions
\link[utils]{methods} and \link[utils]{.S3methods}}

\item{i}{subset criteria, see examples}
}
\value{
Function \link{[.MethodsFunction} returns a \code{'MethodsFunction'} object.
}
\description{
..
}
\details{
..
}
\note{
I am still playing with \link[utils]{getS3method} and \link[methods]{getMethod}

Not sure what is \link[methods]{getClass}
}
\examples{
library(Matrix)
(x1 = .S3methods(head))
x1[1:3]

(x2 = methods(head))
x2[!isS4]
x2[isS4]

# developer's playground
utils::getS3method(f = 'head', class = 'data.frame') |> class() # 'function'
methods::getMethod(f = 'head', signature = 'Matrix') |> class() # 'MethodDefinition'
}
