% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDESCRIPTION.R
\name{updateDESCRIPTION}
\alias{updateDESCRIPTION}
\title{Update DESCRIPTION File}
\usage{
updateDESCRIPTION(pkg)
}
\arguments{
\item{pkg}{\link[base]{character} scalar, directory of the package}
}
\description{
Update DESCRIPTION file
}
\details{
Update the DESCRIPTION file, so that

\describe{

\item{\code{Version}}{is the next version on CRAN, via \link{nextCRANversion}}

\item{\code{Date}}{is today (via \link[base]{Sys.Date})}

}
}
