% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennS3.R
\name{vennS3}
\alias{vennS3}
\title{\link{vennS3}}
\usage{
vennS3(pkg, mc.cores = getOption("cores"), ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar}

\item{mc.cores}{parameter of function \link[parallel]{mclapply}}

\item{...}{additional parameters, currently of no use}
}
\description{
\link{vennS3}
}
\examples{
\dontrun{
vennS3('base')
vennS3('stats')
vennS3('graphics')
vennS3('utils')
vennS3('tools')

vennS3('MASS')
vennS3('nlme')
vennS3('survival')
vennS3('caret')
vennS3('ggplot2')
vennS3('plotly')

# vennS3('spatstat.geom') # base::local() definition causes error..
}


}
\keyword{internal}
