% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{working_package_dependencies}
\alias{working_package_dependencies}
\alias{working_package_dependencies_ext}
\alias{package_dependencies_ext}
\title{Working Package Dependencies}
\usage{
working_package_dependencies(pkg = ".", recursive = TRUE, ...)

working_package_dependencies_ext(...)

package_dependencies_ext(..., recursive = TRUE)
}
\arguments{
\item{pkg}{\link[base]{character} scalar or \link[base]{vector}}

\item{recursive}{\link[base]{logical} scalar, see \link[tools]{package_dependencies}}

\item{...}{additional parameters of \link[tools]{package_dependencies}}
}
\description{
Similar to \link[tools]{package_dependencies}, but for
working packages (i.e., directories on hard drive to be loaded by \link[pkgload]{load_all})
}
\examples{
if (FALSE) {
pkg = c('.', '../ThomasJeffersonUniv', '../DanielBiostatistics10th', '../readData',
  '../TukeyGH77')
working_package_dependencies(pkg, recursive = FALSE)
working_package_dependencies(pkg, recursive = TRUE)
working_package_dependencies_ext(pkg, recursive = TRUE)
working_package_dependencies_date(pkg)
}


}
