% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{package_dependencies_date}
\alias{package_dependencies_date}
\alias{working_package_dependencies_date}
\title{Last Update Dates of Packages Dependencies, in Reverse Order}
\usage{
package_dependencies_date(...)

working_package_dependencies_date(...)
}
\arguments{
\item{...}{parameters of functions \link{working_package_dependencies} and \link[tools]{package_dependencies}}
}
\description{
..
}
\examples{
\dontrun{# needs to set CRAN mirror when ?devtools::check
package_dependencies_date(c('MASS', 'survival', 'nlme'), recursive = TRUE)}
}
