% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{sort_packageDate_}
\alias{sort_packageDate_}
\title{Last Update Dates of Packages, in Reverse Order}
\usage{
sort_packageDate_(pkg, decreasing = TRUE, ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar or \link[base]{vector},
packages on CRAN}

\item{decreasing}{\link[base]{logical} scalar, default \code{TRUE}.
See \link[base]{order}}

\item{...}{additional parameters of \link[utils]{packageDate}}
}
\value{
Function \link{sort_packageDate_} returns a \link[base]{Date} scalar or \link[base]{vector}.
}
\description{
..
}
\keyword{internal}
