% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennS3.R
\name{vennS3}
\alias{vennS3}
\title{Venn Graph of \code{S3} Generic & Methods}
\usage{
vennS3(pkg, mc.cores = getOption("cores"), ...)
}
\arguments{
\item{pkg}{\link[base]{character} scalar, name of package}

\item{mc.cores}{parameter of function \link[parallel]{mclapply}}

\item{...}{additional parameters, currently of no use}
}
\description{
Venn Graph of \code{S3} Generic & Methods
}
\examples{
\dontrun{
vennS3('base')
vennS3('stats')
vennS3('graphics')
vennS3('utils')
vennS3('tools')

vennS3('MASS')
vennS3('nlme')
vennS3('mgcv')
vennS3('survival')
vennS3('caret')
vennS3('ggplot2')
vennS3('plotly')
vennS3('R6')
vennS3('S7')

vennS3('groupedHyperframe')

# vennS3('spatstat.geom') # base::local() definition causes error..
}


}
\keyword{internal}
