% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_package_dependencies.R
\name{working_package_dependencies}
\alias{working_package_dependencies}
\title{Working Package Dependencies}
\usage{
working_package_dependencies(
  path = ".",
  recursive = TRUE,
  vanilla.rm = TRUE,
  ...
)
}
\arguments{
\item{path}{\link[base]{character} scalar}

\item{recursive}{\link[base]{logical} scalar, see \link[tools]{package_dependencies}}

\item{vanilla.rm}{\link[base]{logical} scalar, whether to remove packages in \link[base]{search} path
of a vanilla R session.  Default \code{TRUE}}

\item{...}{additional parameters of \link[tools]{package_dependencies}}
}
\description{
Similar to \link[tools]{package_dependencies}, but for
working packages (i.e., directories on hard drive to be loaded by function \link[pkgload]{load_all}).
}
\note{
I do not know how to make function \link[tools]{package_dependencies} work with packages on Github!
}
\examples{
if (FALSE) {
working_package_dependencies('.', recursive = FALSE)
working_package_dependencies('.', recursive = TRUE)
working_package_dependencies('.', recursive = TRUE, vanilla.rm = FALSE)
working_package_dependencies_date('.')

working_package_dependencies('../DanielBiostatistics10th', recursive = FALSE)
working_package_dependencies('../DanielBiostatistics10th', recursive = TRUE)
working_package_dependencies('../DanielBiostatistics10th', recursive = TRUE, vanilla.rm = FALSE)
working_package_dependencies_date('../DanielBiostatistics10th')
}
}
\keyword{internal}
